compileエラー解消

+++ srcs/http/Request.cpp
@@ -278,7 +278,7 @@ void Request::ParseChunkData(SocketBuff &buffer_) {
 
 // 型変えたほうが綺麗そう
 void Request::ParseContentLengthBody(SocketBuff &buffer_) {
-  if (static_cast<long>(buffer_.GetBuffSize()) >= total_body_size_) {
+  if (buffer_.GetBuffSize() >= total_body_size_) {
     message_.body = buffer_.GetAndErase(total_body_size_);
     parse_status_ = COMPLETE;
   } else {
